better types

+++ src/parse/LineNumberingReader.jl
@@ -49,8 +49,8 @@ column(s::LineNumberingReader) = position(s) - s.lines[line(s)] + 1
 linecol(s::LineNumberingReader) = LineCol(line(s), column(s))
 
 immutable LineCol
-  line::Int64
-  column::Int64
+  line::Int
+  column::Int
 end
 
 Base.isless(x::LineCol, y::LineCol) =
